// src/models/pockemon.m.ts
import { type } from "os";

export type Pokemon = {
    id: string;
    name: string;
    imggif: string;
    imglarge: string;
    imgnormal: string;
    total: string;
    hp: string;
    attack: string;
    defense: string;
    sp_atk: string;
    sp_def: string;
    speed: string;
    type: Array<any>;
}